+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ## [Unreleased]
 
 - Implemented importance output
+- Implemented beta distribution
 
 
 ## [v0.3.2] Python 3.12 passing (2025-05-12)